Make all plugins support both csound 6 and csound 7
The most important work has been to determine what changed within the
csound API and how can that be supported while having one codebase.

A brief abstract of the work:

The OENTRY struct has changed. The "thread" attribute is gone and
the callback functions have been renamed to "init" and "perf". Instead
of an audio callback which was never used, a "deinit" function is added
and RegisterDeinitCallback is gone.

Apparently each instrument/opcode can define a local samplerate. This
does not seem implemented yet, but there is the concept of a local samplerate
and it is not possible to get the global samplerate.

- new: new cmake variable, APIVERSION, which determines for which api we build
  This could be detected but it is better to make it explicit to avoid
  mistakes. Depending on the APIVERSION, two defines are added: CSOUNDAPI6
  is defined in csound 6, CSOUNDAPI7 in csound 7. We avoid to follow
  csound's own api versioning, which just mirrors its version
- new: compatibility functions defined in _common.h.